refactor: [arieftb] Atoms Level Dir
###### Changed

- from Atomics to Atoms
