+++ src/presentation/molecules/input/InputField.jsx
@@ -1,6 +1,6 @@
-import { SmallText } from '../../atomics/text/Text.jsx';
+import { SmallText } from '../../atoms/text/Text.jsx';
 import PropTypes from 'prop-types';
-import Input from '../../atomics/input/Input.jsx';
+import Input from '../../atoms/input/Input.jsx';
 import styles from './InputField.module.css';
 
 const InputField = ({
