+++ src/presentation/organisms/form/SignUpForm.jsx
@@ -1,11 +1,11 @@
-import { H4 } from '../../atomics/text/Heading.jsx';
+import { H4 } from '../../atoms/text/Heading.jsx';
 import styles from './SignUpForm.module.css';
 import TextInputField from '../../molecules/input/TextInputField.jsx';
 import EmailInputField from '../../molecules/input/EmailInputField.jsx';
 import PasswordInputField from '../../molecules/input/PasswordInputField.jsx';
-import Button from '../../atomics/button/Button.jsx';
+import Button from '../../atoms/button/Button.jsx';
 import { useState } from 'react';
-import { SmallText } from '../../atomics/text/Text.jsx';
+import { SmallText } from '../../atoms/text/Text.jsx';
 
 const SignUpForm = ({ onSubmit, isSubmitting, successMessage, errorMessage }) => {
   const [formData, setFormData] = useState({
